+++ src/test/mocks/vscode.ts
@@ -23,3 +23,41 @@ export const window = {
         dispose: () => { }
     })
 };
+
+export class Position {
+    public readonly line: number;
+    public readonly character: number;
+
+    constructor(line: number, character: number) {
+        this.line = line;
+        this.character = character;
+    }
+}
+
+export class Range {
+    public readonly start: Position;
+    public readonly end: Position;
+
+    constructor(start: Position, end: Position) {
+        this.start = start;
+        this.end = end;
+    }
+
+    public with(start: Position = this.start, end: Position = this.end): Range {
+        return new Range(start, end);
+    }
+}
+
+export class TextEdit {
+    public readonly range: Range;
+    public readonly newText: string;
+
+    private constructor(range: Range, newText: string) {
+        this.range = range;
+        this.newText = newText;
+    }
+
+    public static replace(range: Range, newText: string): TextEdit {
+        return new TextEdit(range, newText);
+    }
+}